Remove LHC functions
Deleted unused LHC functions, leaving only Rejection functions.
Renamed LHC.h and LHC.cc in Rej.h and Rej.cc.
